fixed bug ending game after first play
removed the "0" in the parenthesis on line 100 and replaced with a whitespace.  This fixed the problem.  However, next game that was automatically Regenerated an old problem of the stacks of money moving around randomly during game play.

also cut and pasted an old version of lines 70 - 84 (by current line numbering) which fixed a traceback.  We did not identify any existing in older code but this ;cut and paste of identical seemed to fix the problem.